requirements.txt
mysql-connector-python==9.0.0
pyinstaller==6.11.1
pyserial==3.5
zebra==0.1.0